requirements.txt
pymobiledevice3>=4.0.0
PyYAML>=6.0
pathlib2>=2.3.0
Pillow>=9.0.0
